pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "setuptools_scm[toml]>=8.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "adorable-cli"
dynamic = ["version"]
description = "A powerful CLI super agent"
authors = [{ name = "LeonEthan", email = "czlbou2012@gmail.com" }]
license = "MIT"
license-files = ["LICENSE"]
readme = "README.md"
dependencies = [
    "agno>=2.2.8",
    "fastapi>=0.110.0",
    "rich>=13.0.0",
    "typer>=0.12.0",
    "openai>=1.0.0",
    "pyyaml>=6.0.0",
    "ddgs<9.6.0",
    "firecrawl-py>=4.3.6",
    "httpx[socks]>=0.28.1",
    "chonkie>=1.3.1",
    "mcp>=1.17.0",
    "sqlalchemy>=2.0.44",
    "prompt-toolkit>=3.0.0",
    "update>=0.0.1",
    "uvicorn>=0.30.0",
    "lancedb>=0.17.0",
    "pandas>=2.0.0",
]
requires-python = ">=3.10"

[project.scripts]
adorable = "adorable_cli.main:main"
ador = "adorable_cli.main:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
extend-exclude = ["**/.venv/**", "**/venv/**", "**/env/**", "**/site-packages/**"]

[tool.black]
line-length = 100

[tool.setuptools_scm]
# Derive version from Git tags (supports vX.Y.Z or X.Y.Z)
# Optional: configure dev/local version scheme (default is fine)
# version_scheme = "guess-next-dev"
# local_scheme = "node-and-date"
